+++ app/models/early_career_payments/slug_sequence.rb
@@ -36,6 +36,7 @@ module EarlyCareerPayments
       "email-verification",
       "provide-mobile-number",
       "mobile-number",
+      "mobile-verification",
       # payment details phase
       "bank-or-building-society",
       "personal-bank-account",